highway_simulation/requirements.txt
gymnasium==0.29.1
matplotlib==3.9.2
numpy==2.1.3
pygame==2.6.1
PyYAML==6.0.2
stable_baselines3==2.3.2
torch==2.5.1
